long long is wrong; use int64_t instead. clusterfusion 91.87 token/s flashinfer 98.77 token/s

+++ include/5090/llama/kernel_batch_sglang.cuh
@@ -47,7 +47,7 @@ __global__ void __cluster_dims__(CLUSTER_SIZE, 1, 1) LlamaDecoderLayerBatchDecod
     half* residual_output,
     half* w_rms_input,   // hidden_dim
     float eps,
-    long long* positions,
+    int64_t* positions,
     float* cos_sin,
     half* k_cache,
     half* v_cache,
